requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
fake-useragent==0.1.10
future==0.16.0
idna==2.6
line-bot-sdk==1.5.0
lxml==4.2.1
mysqlclient==1.3.12
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
